+++ src/product.ts
@@ -1,0 +1,5 @@
+export const productName = "ReleaseLens for Azure DevOps";
+export const marketplaceUrl =
+  "https://marketplace.visualstudio.com/items?itemName=matancohenmsft.fe-ninja-tools";
+export const repositoryUrl =
+  "https://github.com/matanelcohen/vscode-ado-build-compare";